+++ brain-react/section_07/src/components/templates/TodoEditTemplate/index.jsx
@@ -3,9 +3,7 @@
  *
  * @package components
  */
-import { useMemo, useState, useCallback } from "react";
-import { useParams, useNavigate } from "react-router-dom";
-import { NAVIGATION_PATH } from "../../../constants/navigations";
+import { useTodoEditTemplate } from "./useTodoEditTemplate";
 import { useTodoContext } from "../../../contexts/TodoContext";
 import { BaseLayout } from "../../organisims/BaseLayout";
 import { InputForm } from "../../atoms/InputForm";
@@ -20,54 +18,12 @@ import styles from "./styles.module.css";
  * @constructor
  */
 export const TodoEditTemplate = () => {
-    const navigate = useNavigate();
     const { originTodoList, updateTodo } = useTodoContext();
-    const { id } = useParams();
 
-    const todo = useMemo(
-        () => originTodoList.find((todo) => String(todo.id) === id),
-        [id, originTodoList]
-    );
-
-    /* local state */
-    const [inputTitle, setInputTitle] = useState(todo?.title || "");
-    const [inputContent, setInputContent] = useState(todo?.content || "");
-
-    /**
-     * 「title」変更処理
-     *
-     * @type {function(*): void}
-     */
-    const handleChangeTitle = useCallback(
-        (e) => setInputTitle(e.target.value),
-        []
-    );
-
-    /**
-     * 「content」変更処理
-     *
-     * @type {function(*): void}
-     */
-    const handleChangeContent = useCallback(
-        (e) => setInputContent(e.target.value),
-        []
-    );
-
-    /**
-     * Todo更新処理
-     *
-     * @type {(function(*): void)|*}
-     */
-    const handleUpdateTodo = useCallback(
-        (e) => {
-            e.preventDefault();
-            if (!!todo?.id && inputTitle !== "" && inputContent !== "") {
-                updateTodo(todo.id, inputTitle, inputContent);
-                navigate(NAVIGATION_PATH.TOP);
-            }
-        },
-        [navigate, todo?.id, inputTitle, inputContent, updateTodo]
-    );
+    const [
+        { todo, inputTitle, inputContent },
+        { handleChangeTitle, handleChangeContent, handleUpdateTodo }
+    ] = useTodoEditTemplate({ originTodoList, updateTodo });
 
     return (
         <BaseLayout title={"TodoEdit"}>
